studykrack.web/legacy_archive/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
google-genai==1.3.0
firebase-admin==6.6.0
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==3.1.1
numpy==1.26.4
faiss-cpu==1.8.0
